Update aptupdater
Made lots of changes. Removed unnecessary time check function and made it just one function (no need to over complicate it you were essentially doing the same thing twice in two different nested functions). I also changed the location of the log files. /usr/bin should be strictly for executables (for example to run this program from the command line it should be saved under /usr/bin), log files and anything else shouldn't be stored here (it is part of $PATH there for should only contain executable programs). I changed the log directory to a more suitable location: /var/lib. I made it so it looks for the directory /var/lib/aptupdater and if it does not exist creates it. I changed all the log file variables "F_1-3" to point to the correct place and also replaced all your absolute paths with just the variables (this makes it simpler to change the location of the logs without having to go through and change every single path. Simplified check logs and check time they both work perfect and have more colors. I also totally simplified delete logs (no reason to make it complex with if statements if the goal is to delete the files anyway). One of the really cool things I added is lolcat which makes apt have rainbow output and looks cool (hackerish) in the terminal. Lolcat is a program like cat but it has rainbow color output. for example: echo "This text is rainbow" | lolcat. I made the script so it detects if the user has lolcat installed and if they do it enables lolcat when running any update (if they dont have lolcat installed it runs without piping to lolcat). To install lolcat simply run: sudo apt install lolcat

Thats about it for now. Overall its less lines and works a bit better (more pretty haha) ;) Enjoy